requirements.txt
python-docx == 1.1.2
openai == 1.52.1
openai == 1.52.1